Moved course search to a new component.
