+++ src/Components/DegreePlanComponents/TableFooter.tsx
@@ -1,15 +1,14 @@
 import React from "react";
-import { Button, Form } from "react-bootstrap";
+import { Button } from "react-bootstrap";
 import { Course } from "../../Interfaces/course";
 import { DegreePlan } from "../../Interfaces/degreePlan";
 import { Semester } from "../../Interfaces/semester";
-import { findCourseByCode } from "./ViewPlanFunctions";
 import { addCourse, clearSem } from "./ViewPlanFunctions";
+import { CourseSearch } from "./CourseSearchComponent";
 
 //Icon imports for buttons
 import { AiOutlineClear } from "react-icons/ai";
 import { RiAddBoxLine } from "react-icons/ri";
-import SelectSearch, { SelectSearchOption } from "react-select-search";
 
 interface footerProps {
     plan: DegreePlan;
@@ -32,39 +31,14 @@ export function TableFooter({
     courses,
     setEditingSem
 }: footerProps): JSX.Element {
-    const sel_search_options: SelectSearchOption[] = courses.map(
-        (course: Course) => {
-            const opt: SelectSearchOption = {
-                name: course.code,
-                value: course.code
-            };
-            return opt;
-        }
-    );
-    function filter_for_sel_search(opts: SelectSearchOption[]) {
-        return (opt: string) =>
-            opts.filter((val: SelectSearchOption) => val.name.startsWith(opt));
-    }
     return (
         <tbody>
             <tr>
                 <td colSpan={1}>
-                    <h5>Search for course:</h5>
-                    <SelectSearch
-                        options={sel_search_options}
-                        search={true}
-                        filterOptions={filter_for_sel_search}
-                        closeOnSelect={false}
-                        onChange={(selectedVal) =>
-                            setAddingCourse(
-                                findCourseByCode(
-                                    courses,
-                                    selectedVal.toString()
-                                )
-                            )
-                        }
-                        placeholder={"ex. CISC 108"}
-                    ></SelectSearch>
+                    <CourseSearch
+                        courses={courses}
+                        setAddingCourse={setAddingCourse}
+                    ></CourseSearch>
                     {/* 
                     <Form.Group controlId="addingCourse">
                         <Form.Label>Pick the Course to be Added:</Form.Label>
